Same as `PointRCNN`, negative values in `point_part_labels` should be masked out. Related links:
- https://github.com/liyufan/OpenPCDet/commit/eba98549b6bc4f28ed88564baf21e13d69decbe1
- https://github.com/open-mmlab/OpenPCDet/issues/1682
- https://github.com/open-mmlab/OpenPCDet/issues/1691

Simply by adding some printing to `pcdet/models/dense_heads/point_head_template.py`:
```py
# pcdet/models/dense_heads/point_head_template.py
def get_part_layer_loss(self, tb_dict=None):
    pos_mask = self.forward_ret_dict['point_cls_labels'] > 0
    pos_normalizer = max(1, (pos_mask > 0).sum().item())
    point_part_labels = self.forward_ret_dict['point_part_labels']
    point_part_preds = self.forward_ret_dict['point_part_preds']
    valid_row_mask = (point_part_labels >= 0).all(dim=-1)
    valid_mask = pos_mask & valid_row_mask
    point_loss_part = F.binary_cross_entropy_with_logits(
        point_part_preds[valid_mask], point_part_labels[valid_mask], reduction='mean'
    )
    print(f'corrected point_loss_part: {point_loss_part}', end=' ----- ')
    point_loss_part = F.binary_cross_entropy(
        torch.sigmoid(point_part_preds), point_part_labels, reduction='none'
    )
    point_loss_part = (point_loss_part.sum(dim=-1) * pos_mask.float()).sum() / (
        3 * pos_normalizer
    )
    print(f'original point_loss_part: {point_loss_part}')
    ...
```
And the output is:
```
(py) lyf@ASUS:~/OpenPCDet/tools$ CUDA_LAUNCH_BLOCKING=1 python train.py --cfg_file cfgs/kitti_models/PartA2.yaml
/home/lyf/OpenPCDet/tools/../pcdet/utils/loss_utils.py:152: FutureWarning: `torch.cuda.amp.custom_fwd(args...)` is deprecated. Please use `torch.amp.custom_fwd(args..., device_type='cuda')` instead.
  @torch.cuda.amp.custom_fwd(cast_inputs=torch.float16)
2025-05-02 17:59:17,835   INFO  **********************Start logging**********************
2025-05-02 17:59:17,835   INFO  CUDA_VISIBLE_DEVICES=ALL
2025-05-02 17:59:17,835   INFO  Training with a single process
2025-05-02 17:59:17,835   INFO  cfg_file         cfgs/kitti_models/PartA2.yaml
2025-05-02 17:59:17,835   INFO  batch_size       4
2025-05-02 17:59:17,835   INFO  epochs           80
2025-05-02 17:59:17,835   INFO  workers          4
2025-05-02 17:59:17,835   INFO  extra_tag        default
2025-05-02 17:59:17,835   INFO  ckpt             None
2025-05-02 17:59:17,835   INFO  pretrained_model None
2025-05-02 17:59:17,835   INFO  launcher         none
2025-05-02 17:59:17,835   INFO  tcp_port         18888
2025-05-02 17:59:17,835   INFO  sync_bn          False
2025-05-02 17:59:17,835   INFO  fix_random_seed  False
2025-05-02 17:59:17,835   INFO  ckpt_save_interval 1
2025-05-02 17:59:17,835   INFO  local_rank       None
2025-05-02 17:59:17,835   INFO  max_ckpt_save_num 30
2025-05-02 17:59:17,835   INFO  merge_all_iters_to_one_epoch False
2025-05-02 17:59:17,835   INFO  set_cfgs         None
2025-05-02 17:59:17,835   INFO  max_waiting_mins 0
2025-05-02 17:59:17,835   INFO  start_epoch      0
2025-05-02 17:59:17,835   INFO  num_epochs_to_eval 0
2025-05-02 17:59:17,835   INFO  save_to_file     False
2025-05-02 17:59:17,835   INFO  use_tqdm_to_record False
2025-05-02 17:59:17,835   INFO  logger_iter_interval 50
2025-05-02 17:59:17,835   INFO  ckpt_save_time_interval 300
2025-05-02 17:59:17,835   INFO  wo_gpu_stat      False
2025-05-02 17:59:17,835   INFO  use_amp          False
2025-05-02 17:59:17,835   INFO  cfg.ROOT_DIR: /home/lyf/OpenPCDet
2025-05-02 17:59:17,835   INFO  cfg.LOCAL_RANK: 0
2025-05-02 17:59:17,835   INFO  cfg.CLASS_NAMES: ['Car', 'Pedestrian', 'Cyclist']
2025-05-02 17:59:17,835   INFO  ----------- DATA_CONFIG -----------
2025-05-02 17:59:17,835   INFO  cfg.DATA_CONFIG.DATASET: KittiDataset
2025-05-02 17:59:17,835   INFO  cfg.DATA_CONFIG.DATA_PATH: ../data/kitti
2025-05-02 17:59:17,835   INFO  cfg.DATA_CONFIG.POINT_CLOUD_RANGE: [0, -40, -3, 70.4, 40, 1]
2025-05-02 17:59:17,835   INFO  ----------- DATA_SPLIT -----------
2025-05-02 17:59:17,835   INFO  cfg.DATA_CONFIG.DATA_SPLIT.train: train
2025-05-02 17:59:17,835   INFO  cfg.DATA_CONFIG.DATA_SPLIT.test: val
2025-05-02 17:59:17,835   INFO  ----------- INFO_PATH -----------
2025-05-02 17:59:17,835   INFO  cfg.DATA_CONFIG.INFO_PATH.train: ['kitti_infos_train.pkl']
2025-05-02 17:59:17,835   INFO  cfg.DATA_CONFIG.INFO_PATH.test: ['kitti_infos_val.pkl']
2025-05-02 17:59:17,835   INFO  cfg.DATA_CONFIG.GET_ITEM_LIST: ['points']
2025-05-02 17:59:17,835   INFO  cfg.DATA_CONFIG.FOV_POINTS_ONLY: True
2025-05-02 17:59:17,835   INFO  ----------- DATA_AUGMENTOR -----------
2025-05-02 17:59:17,835   INFO  cfg.DATA_CONFIG.DATA_AUGMENTOR.DISABLE_AUG_LIST: ['placeholder']
2025-05-02 17:59:17,835   INFO  cfg.DATA_CONFIG.DATA_AUGMENTOR.AUG_CONFIG_LIST: [{'NAME': 'gt_sampling', 'USE_ROAD_PLANE': True, 'DB_INFO_PATH': ['kitti_dbinfos_train.pkl'], 'PREPARE': {'filter_by_min_points': ['Car:5', 'Pedestrian:5', 'Cyclist:5'], 'filter_by_difficulty': [-1]}, 'SAMPLE_GROUPS': ['Car:20', 'Pedestrian:15', 'Cyclist:15'], 'NUM_POINT_FEATURES': 4, 'DATABASE_WITH_FAKELIDAR': False, 'REMOVE_EXTRA_WIDTH': [0.0, 0.0, 0.0], 'LIMIT_WHOLE_SCENE': True}, {'NAME': 'random_world_flip', 'ALONG_AXIS_LIST': ['x']}, {'NAME': 'random_world_rotation', 'WORLD_ROT_ANGLE': [-0.78539816, 0.78539816]}, {'NAME': 'random_world_scaling', 'WORLD_SCALE_RANGE': [0.95, 1.05]}]
2025-05-02 17:59:17,835   INFO  ----------- POINT_FEATURE_ENCODING -----------
2025-05-02 17:59:17,835   INFO  cfg.DATA_CONFIG.POINT_FEATURE_ENCODING.encoding_type: absolute_coordinates_encoding
2025-05-02 17:59:17,835   INFO  cfg.DATA_CONFIG.POINT_FEATURE_ENCODING.used_feature_list: ['x', 'y', 'z', 'intensity']
2025-05-02 17:59:17,835   INFO  cfg.DATA_CONFIG.POINT_FEATURE_ENCODING.src_feature_list: ['x', 'y', 'z', 'intensity']
2025-05-02 17:59:17,835   INFO  cfg.DATA_CONFIG.DATA_PROCESSOR: [{'NAME': 'mask_points_and_boxes_outside_range', 'REMOVE_OUTSIDE_BOXES': True}, {'NAME': 'shuffle_points', 'SHUFFLE_ENABLED': {'train': True, 'test': False}}, {'NAME': 'transform_points_to_voxels', 'VOXEL_SIZE': [0.05, 0.05, 0.1], 'MAX_POINTS_PER_VOXEL': 5, 'MAX_NUMBER_OF_VOXELS': {'train': 16000, 'test': 40000}}]
2025-05-02 17:59:17,835   INFO  cfg.DATA_CONFIG._BASE_CONFIG_: cfgs/dataset_configs/kitti_dataset.yaml
2025-05-02 17:59:17,835   INFO  ----------- MODEL -----------
2025-05-02 17:59:17,835   INFO  cfg.MODEL.NAME: PartA2Net
2025-05-02 17:59:17,835   INFO  ----------- VFE -----------
2025-05-02 17:59:17,835   INFO  cfg.MODEL.VFE.NAME: MeanVFE
2025-05-02 17:59:17,835   INFO  ----------- BACKBONE_3D -----------
2025-05-02 17:59:17,835   INFO  cfg.MODEL.BACKBONE_3D.NAME: UNetV2
2025-05-02 17:59:17,835   INFO  ----------- MAP_TO_BEV -----------
2025-05-02 17:59:17,835   INFO  cfg.MODEL.MAP_TO_BEV.NAME: HeightCompression
2025-05-02 17:59:17,835   INFO  cfg.MODEL.MAP_TO_BEV.NUM_BEV_FEATURES: 256
2025-05-02 17:59:17,835   INFO  ----------- BACKBONE_2D -----------
2025-05-02 17:59:17,835   INFO  cfg.MODEL.BACKBONE_2D.NAME: BaseBEVBackbone
2025-05-02 17:59:17,835   INFO  cfg.MODEL.BACKBONE_2D.LAYER_NUMS: [5, 5]
2025-05-02 17:59:17,835   INFO  cfg.MODEL.BACKBONE_2D.LAYER_STRIDES: [1, 2]
2025-05-02 17:59:17,835   INFO  cfg.MODEL.BACKBONE_2D.NUM_FILTERS: [128, 256]
2025-05-02 17:59:17,835   INFO  cfg.MODEL.BACKBONE_2D.UPSAMPLE_STRIDES: [1, 2]
2025-05-02 17:59:17,835   INFO  cfg.MODEL.BACKBONE_2D.NUM_UPSAMPLE_FILTERS: [256, 256]
2025-05-02 17:59:17,835   INFO  ----------- DENSE_HEAD -----------
2025-05-02 17:59:17,835   INFO  cfg.MODEL.DENSE_HEAD.NAME: AnchorHeadSingle
2025-05-02 17:59:17,835   INFO  cfg.MODEL.DENSE_HEAD.CLASS_AGNOSTIC: False
2025-05-02 17:59:17,835   INFO  cfg.MODEL.DENSE_HEAD.USE_DIRECTION_CLASSIFIER: True
2025-05-02 17:59:17,835   INFO  cfg.MODEL.DENSE_HEAD.DIR_OFFSET: 0.78539
2025-05-02 17:59:17,835   INFO  cfg.MODEL.DENSE_HEAD.DIR_LIMIT_OFFSET: 0.0
2025-05-02 17:59:17,835   INFO  cfg.MODEL.DENSE_HEAD.NUM_DIR_BINS: 2
2025-05-02 17:59:17,835   INFO  cfg.MODEL.DENSE_HEAD.ANCHOR_GENERATOR_CONFIG: [{'class_name': 'Car', 'anchor_sizes': [[3.9, 1.6, 1.56]], 'anchor_rotations': [0, 1.57], 'anchor_bottom_heights': [-1.78], 'align_center': False, 'feature_map_stride': 8, 'matched_threshold': 0.6, 'unmatched_threshold': 0.45}, {'class_name': 'Pedestrian', 'anchor_sizes': [[0.8, 0.6, 1.73]], 'anchor_rotations': [0, 1.57], 'anchor_bottom_heights': [-1.78], 'align_center': False, 'feature_map_stride': 8, 'matched_threshold': 0.5, 'unmatched_threshold': 0.35}, {'class_name': 'Cyclist', 'anchor_sizes': [[1.76, 0.6, 1.73]], 'anchor_rotations': [0, 1.57], 'anchor_bottom_heights': [-1.78], 'align_center': False, 'feature_map_stride': 8, 'matched_threshold': 0.5, 'unmatched_threshold': 0.35}]
2025-05-02 17:59:17,835   INFO  ----------- TARGET_ASSIGNER_CONFIG -----------
2025-05-02 17:59:17,835   INFO  cfg.MODEL.DENSE_HEAD.TARGET_ASSIGNER_CONFIG.NAME: AxisAlignedTargetAssigner
2025-05-02 17:59:17,835   INFO  cfg.MODEL.DENSE_HEAD.TARGET_ASSIGNER_CONFIG.POS_FRACTION: -1.0
2025-05-02 17:59:17,835   INFO  cfg.MODEL.DENSE_HEAD.TARGET_ASSIGNER_CONFIG.SAMPLE_SIZE: 512
2025-05-02 17:59:17,835   INFO  cfg.MODEL.DENSE_HEAD.TARGET_ASSIGNER_CONFIG.NORM_BY_NUM_EXAMPLES: False
2025-05-02 17:59:17,835   INFO  cfg.MODEL.DENSE_HEAD.TARGET_ASSIGNER_CONFIG.MATCH_HEIGHT: False
2025-05-02 17:59:17,835   INFO  cfg.MODEL.DENSE_HEAD.TARGET_ASSIGNER_CONFIG.BOX_CODER: ResidualCoder
2025-05-02 17:59:17,835   INFO  ----------- LOSS_CONFIG -----------
2025-05-02 17:59:17,835   INFO  ----------- LOSS_WEIGHTS -----------
2025-05-02 17:59:17,835   INFO  cfg.MODEL.DENSE_HEAD.LOSS_CONFIG.LOSS_WEIGHTS.cls_weight: 1.0
2025-05-02 17:59:17,835   INFO  cfg.MODEL.DENSE_HEAD.LOSS_CONFIG.LOSS_WEIGHTS.loc_weight: 2.0
2025-05-02 17:59:17,835   INFO  cfg.MODEL.DENSE_HEAD.LOSS_CONFIG.LOSS_WEIGHTS.dir_weight: 0.2
2025-05-02 17:59:17,835   INFO  cfg.MODEL.DENSE_HEAD.LOSS_CONFIG.LOSS_WEIGHTS.code_weights: [1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0]
2025-05-02 17:59:17,836   INFO  ----------- POINT_HEAD -----------
2025-05-02 17:59:17,836   INFO  cfg.MODEL.POINT_HEAD.NAME: PointIntraPartOffsetHead
2025-05-02 17:59:17,836   INFO  cfg.MODEL.POINT_HEAD.CLS_FC: []
2025-05-02 17:59:17,836   INFO  cfg.MODEL.POINT_HEAD.PART_FC: []
2025-05-02 17:59:17,836   INFO  cfg.MODEL.POINT_HEAD.CLASS_AGNOSTIC: True
2025-05-02 17:59:17,836   INFO  ----------- TARGET_CONFIG -----------
2025-05-02 17:59:17,836   INFO  cfg.MODEL.POINT_HEAD.TARGET_CONFIG.GT_EXTRA_WIDTH: [0.2, 0.2, 0.2]
2025-05-02 17:59:17,836   INFO  ----------- LOSS_CONFIG -----------
2025-05-02 17:59:17,836   INFO  cfg.MODEL.POINT_HEAD.LOSS_CONFIG.LOSS_REG: smooth-l1
2025-05-02 17:59:17,836   INFO  ----------- LOSS_WEIGHTS -----------
2025-05-02 17:59:17,836   INFO  cfg.MODEL.POINT_HEAD.LOSS_CONFIG.LOSS_WEIGHTS.point_cls_weight: 1.0
2025-05-02 17:59:17,836   INFO  cfg.MODEL.POINT_HEAD.LOSS_CONFIG.LOSS_WEIGHTS.point_part_weight: 1.0
2025-05-02 17:59:17,836   INFO  ----------- ROI_HEAD -----------
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.NAME: PartA2FCHead
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.CLASS_AGNOSTIC: True
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.SHARED_FC: [256, 256, 256]
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.CLS_FC: [256, 256]
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.REG_FC: [256, 256]
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.DP_RATIO: 0.3
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.SEG_MASK_SCORE_THRESH: 0.3
2025-05-02 17:59:17,836   INFO  ----------- NMS_CONFIG -----------
2025-05-02 17:59:17,836   INFO  ----------- TRAIN -----------
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.NMS_CONFIG.TRAIN.NMS_TYPE: nms_gpu
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.NMS_CONFIG.TRAIN.MULTI_CLASSES_NMS: False
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.NMS_CONFIG.TRAIN.NMS_PRE_MAXSIZE: 9000
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.NMS_CONFIG.TRAIN.NMS_POST_MAXSIZE: 512
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.NMS_CONFIG.TRAIN.NMS_THRESH: 0.8
2025-05-02 17:59:17,836   INFO  ----------- TEST -----------
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.NMS_CONFIG.TEST.NMS_TYPE: nms_gpu
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.NMS_CONFIG.TEST.MULTI_CLASSES_NMS: False
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.NMS_CONFIG.TEST.NMS_PRE_MAXSIZE: 1024
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.NMS_CONFIG.TEST.NMS_POST_MAXSIZE: 100
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.NMS_CONFIG.TEST.NMS_THRESH: 0.7
2025-05-02 17:59:17,836   INFO  ----------- ROI_AWARE_POOL -----------
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.ROI_AWARE_POOL.POOL_SIZE: 12
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.ROI_AWARE_POOL.NUM_FEATURES: 128
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.ROI_AWARE_POOL.MAX_POINTS_PER_VOXEL: 128
2025-05-02 17:59:17,836   INFO  ----------- TARGET_CONFIG -----------
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.TARGET_CONFIG.BOX_CODER: ResidualCoder
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.TARGET_CONFIG.ROI_PER_IMAGE: 128
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.TARGET_CONFIG.FG_RATIO: 0.5
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.TARGET_CONFIG.SAMPLE_ROI_BY_EACH_CLASS: True
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.TARGET_CONFIG.CLS_SCORE_TYPE: roi_iou
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.TARGET_CONFIG.CLS_FG_THRESH: 0.75
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.TARGET_CONFIG.CLS_BG_THRESH: 0.25
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.TARGET_CONFIG.CLS_BG_THRESH_LO: 0.1
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.TARGET_CONFIG.HARD_BG_RATIO: 0.8
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.TARGET_CONFIG.REG_FG_THRESH: 0.65
2025-05-02 17:59:17,836   INFO  ----------- LOSS_CONFIG -----------
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.LOSS_CONFIG.CLS_LOSS: BinaryCrossEntropy
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.LOSS_CONFIG.REG_LOSS: smooth-l1
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.LOSS_CONFIG.CORNER_LOSS_REGULARIZATION: True
2025-05-02 17:59:17,836   INFO  ----------- LOSS_WEIGHTS -----------
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.LOSS_CONFIG.LOSS_WEIGHTS.rcnn_cls_weight: 1.0
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.LOSS_CONFIG.LOSS_WEIGHTS.rcnn_reg_weight: 1.0
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.LOSS_CONFIG.LOSS_WEIGHTS.rcnn_corner_weight: 1.0
2025-05-02 17:59:17,836   INFO  cfg.MODEL.ROI_HEAD.LOSS_CONFIG.LOSS_WEIGHTS.code_weights: [1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0]
2025-05-02 17:59:17,836   INFO  ----------- POST_PROCESSING -----------
2025-05-02 17:59:17,836   INFO  cfg.MODEL.POST_PROCESSING.RECALL_THRESH_LIST: [0.3, 0.5, 0.7]
2025-05-02 17:59:17,836   INFO  cfg.MODEL.POST_PROCESSING.SCORE_THRESH: 0.1
2025-05-02 17:59:17,836   INFO  cfg.MODEL.POST_PROCESSING.OUTPUT_RAW_SCORE: False
2025-05-02 17:59:17,836   INFO  cfg.MODEL.POST_PROCESSING.EVAL_METRIC: kitti
2025-05-02 17:59:17,836   INFO  ----------- NMS_CONFIG -----------
2025-05-02 17:59:17,836   INFO  cfg.MODEL.POST_PROCESSING.NMS_CONFIG.MULTI_CLASSES_NMS: False
2025-05-02 17:59:17,836   INFO  cfg.MODEL.POST_PROCESSING.NMS_CONFIG.NMS_TYPE: nms_gpu
2025-05-02 17:59:17,836   INFO  cfg.MODEL.POST_PROCESSING.NMS_CONFIG.NMS_THRESH: 0.1
2025-05-02 17:59:17,836   INFO  cfg.MODEL.POST_PROCESSING.NMS_CONFIG.NMS_PRE_MAXSIZE: 4096
2025-05-02 17:59:17,836   INFO  cfg.MODEL.POST_PROCESSING.NMS_CONFIG.NMS_POST_MAXSIZE: 500
2025-05-02 17:59:17,836   INFO  ----------- OPTIMIZATION -----------
2025-05-02 17:59:17,836   INFO  cfg.OPTIMIZATION.BATCH_SIZE_PER_GPU: 4
2025-05-02 17:59:17,836   INFO  cfg.OPTIMIZATION.NUM_EPOCHS: 80
2025-05-02 17:59:17,836   INFO  cfg.OPTIMIZATION.OPTIMIZER: adam_onecycle
2025-05-02 17:59:17,836   INFO  cfg.OPTIMIZATION.LR: 0.01
2025-05-02 17:59:17,836   INFO  cfg.OPTIMIZATION.WEIGHT_DECAY: 0.01
2025-05-02 17:59:17,836   INFO  cfg.OPTIMIZATION.MOMENTUM: 0.9
2025-05-02 17:59:17,836   INFO  cfg.OPTIMIZATION.MOMS: [0.95, 0.85]
2025-05-02 17:59:17,836   INFO  cfg.OPTIMIZATION.PCT_START: 0.4
2025-05-02 17:59:17,836   INFO  cfg.OPTIMIZATION.DIV_FACTOR: 10
2025-05-02 17:59:17,836   INFO  cfg.OPTIMIZATION.DECAY_STEP_LIST: [35, 45]
2025-05-02 17:59:17,836   INFO  cfg.OPTIMIZATION.LR_DECAY: 0.1
2025-05-02 17:59:17,836   INFO  cfg.OPTIMIZATION.LR_CLIP: 1e-07
2025-05-02 17:59:17,836   INFO  cfg.OPTIMIZATION.LR_WARMUP: False
2025-05-02 17:59:17,836   INFO  cfg.OPTIMIZATION.WARMUP_EPOCH: 1
2025-05-02 17:59:17,836   INFO  cfg.OPTIMIZATION.GRAD_NORM_CLIP: 10
2025-05-02 17:59:17,836   INFO  cfg.TAG: PartA2
2025-05-02 17:59:17,836   INFO  cfg.EXP_GROUP_PATH: kitti_models
2025-05-02 17:59:17,838   INFO  ----------- Create dataloader & network & optimizer -----------
2025-05-02 17:59:17,887   INFO  Database filter by min points Car: 14357 => 13532
2025-05-02 17:59:17,887   INFO  Database filter by min points Pedestrian: 2207 => 2168
2025-05-02 17:59:17,887   INFO  Database filter by min points Cyclist: 734 => 705
2025-05-02 17:59:17,887   INFO  Database filter by difficulty Car: 13532 => 10759
2025-05-02 17:59:17,887   INFO  Database filter by difficulty Pedestrian: 2168 => 2075
2025-05-02 17:59:17,887   INFO  Database filter by difficulty Cyclist: 705 => 581
2025-05-02 17:59:17,890   INFO  Loading KITTI dataset
2025-05-02 17:59:17,932   INFO  Total samples for KITTI dataset: 3712
/home/lyf/anaconda3/envs/py/lib/python3.13/site-packages/torch/functional.py:554: UserWarning: torch.meshgrid: in an upcoming release, it will be required to pass the indexing argument. (Triggered internally at /pytorch/aten/src/ATen/native/TensorShape.cpp:4314.)
  return _VF.meshgrid(tensors, **kwargs)  # type: ignore[attr-defined]
2025-05-02 17:59:18,433   INFO  ----------- Model PartA2Net created, param count: 63814772 -----------
2025-05-02 17:59:18,433   INFO  PartA2Net(
  (vfe): MeanVFE()
  (backbone_3d): UNetV2(
    (conv_input): SparseSequential(
      (0): SubMConv3d(4, 16, kernel_size=[3, 3, 3], stride=[1, 1, 1], padding=[1, 1, 1], dilation=[1, 1, 1], output_padding=[0, 0, 0], bias=False, algo=ConvAlgo.MaskImplicitGemm)
      (1): BatchNorm1d(16, eps=0.001, momentum=0.01, affine=True, track_running_stats=True)
      (2): ReLU()
    )
    (conv1): SparseSequential(
      (0): SparseSequential(
        (0): SubMConv3d(16, 16, kernel_size=[3, 3, 3], stride=[1, 1, 1], padding=[0, 0, 0], dilation=[1, 1, 1], output_padding=[0, 0, 0], bias=False, algo=ConvAlgo.MaskImplicitGemm)
        (1): BatchNorm1d(16, eps=0.001, momentum=0.01, affine=True, track_running_stats=True)
        (2): ReLU()
      )
    )
    (conv2): SparseSequential(
      (0): SparseSequential(
        (0): SparseConv3d(16, 32, kernel_size=[3, 3, 3], stride=[2, 2, 2], padding=[1, 1, 1], dilation=[1, 1, 1], output_padding=[0, 0, 0], bias=False, algo=ConvAlgo.MaskImplicitGemm)
        (1): BatchNorm1d(32, eps=0.001, momentum=0.01, affine=True, track_running_stats=True)
        (2): ReLU()
      )
      (1): SparseSequential(
        (0): SubMConv3d(32, 32, kernel_size=[3, 3, 3], stride=[1, 1, 1], padding=[0, 0, 0], dilation=[1, 1, 1], output_padding=[0, 0, 0], bias=False, algo=ConvAlgo.MaskImplicitGemm)
        (1): BatchNorm1d(32, eps=0.001, momentum=0.01, affine=True, track_running_stats=True)
        (2): ReLU()
      )
      (2): SparseSequential(
        (0): SubMConv3d(32, 32, kernel_size=[3, 3, 3], stride=[1, 1, 1], padding=[0, 0, 0], dilation=[1, 1, 1], output_padding=[0, 0, 0], bias=False, algo=ConvAlgo.MaskImplicitGemm)
        (1): BatchNorm1d(32, eps=0.001, momentum=0.01, affine=True, track_running_stats=True)
        (2): ReLU()
      )
    )
    (conv3): SparseSequential(
      (0): SparseSequential(
…